# Fix performance regression