--- a/cabal.project.nothunks.ghc-9.2.4
+++ b/cabal.project.nothunks.ghc-9.2.4
@@ -1,0 +1,1 @@
+cabal.project.nothunks