--- a/scripts/vehicle-debug
+++ b/scripts/vehicle-debug
@@ -1,2 +1,0 @@
-#!/bin/sh
-cabal --project-file=cabal.project.debug run exe:vehicle -- $@