--- a/scripts/vehicle-ghc-debug
+++ b/scripts/vehicle-ghc-debug
@@ -1,0 +1,2 @@
+#!/bin/sh
+cabal --project-file=cabal.project.ghc-debug run vehicle:exe:vehicle -- $@