--- a/scripts/vehicle-nothunks
+++ b/scripts/vehicle-nothunks
@@ -1,0 +1,2 @@
+#!/bin/sh
+cabal --project-file=cabal.project.nothunks run vehicle:exe:vehicle -- $@