--- a/vehicle-syntax/src/Vehicle/Syntax/AST.hs
+++ b/vehicle-syntax/src/Vehicle/Syntax/AST.hs
@@ -8,9 +8,15 @@ import Vehicle.Syntax.AST.Binder as X
 import Vehicle.Syntax.AST.Builtin as X
 import Vehicle.Syntax.AST.Decl as X
 import Vehicle.Syntax.AST.Expr as X
+import Vehicle.Syntax.AST.Instances.NoThunks ()
 import Vehicle.Syntax.AST.Meta as X
 import Vehicle.Syntax.AST.Name as X
 import Vehicle.Syntax.AST.Prog as X
-import Vehicle.Syntax.AST.Provenance as X
+import Vehicle.Syntax.AST.Provenance as X hiding
+  ( Origin,
+    Owner,
+    Position,
+    Range,
+  )
 import Vehicle.Syntax.AST.Relevance as X
 import Vehicle.Syntax.AST.Visibility as X