--- a/vehicle-syntax/src/Vehicle/Syntax/AST/Arg.hs
+++ b/vehicle-syntax/src/Vehicle/Syntax/AST/Arg.hs
@@ -3,7 +3,6 @@ module Vehicle.Syntax.AST.Arg where
 import Control.DeepSeq (NFData)
 import Data.Aeson (FromJSON, ToJSON)
 import GHC.Generics (Generic)
-
 import Vehicle.Syntax.AST.Binder
 import Vehicle.Syntax.AST.Provenance
 import Vehicle.Syntax.AST.Relevance
@@ -15,18 +14,21 @@ import Vehicle.Syntax.AST.Visibility
 -- | An argument to a function, parameterised by the type of expression it
 -- stores.
 data GenericArg expr = Arg
-  { argProvenance :: Provenance
-    -- ^ Has the argument been auto-inserted by the type-checker?
-  , argVisibility :: Visibility
-    -- ^ The visibility of the argument
-  , argRelevance  :: Relevance
-    -- ^ The relevancy of the argument
-  , argExpr       :: expr
-    -- ^ The argument expression
-  } deriving (Eq, Show, Functor, Foldable, Traversable, Generic)
+  { -- | Has the argument been auto-inserted by the type-checker?
+    argProvenance :: Provenance,
+    -- | The visibility of the argument
+    argVisibility :: Visibility,
+    -- | The relevancy of the argument
+    argRelevance :: Relevance,
+    -- | The argument expression
+    argExpr :: expr
+  }
+  deriving (Eq, Show, Functor, Foldable, Traversable, Generic)
 
-instance NFData   expr => NFData   (GenericArg expr)
-instance ToJSON   expr => ToJSON   (GenericArg expr)
+instance NFData expr => NFData (GenericArg expr)
+
+instance ToJSON expr => ToJSON (GenericArg expr)
+
 instance FromJSON expr => FromJSON (GenericArg expr)
 
 instance HasProvenance (GenericArg expr) where
@@ -68,13 +70,14 @@ unpairArg (Arg p v r (x, y)) = (Arg p v r x, y)
 replaceArgExpr :: expr1 -> GenericArg expr2 -> GenericArg expr1
 replaceArgExpr e = fmap (const e)
 
-traverseNonInstanceArgExpr :: Monad m
-                           => (expr -> m expr)
-                           -> GenericArg expr
-                           -> m (GenericArg expr)
+traverseNonInstanceArgExpr ::
+  Monad m =>
+  (expr -> m expr) ->
+  GenericArg expr ->
+  m (GenericArg expr)
 traverseNonInstanceArgExpr f arg
   | isInstance arg = return arg
-  | otherwise      = traverse f arg
+  | otherwise = traverse f arg
 
 argFromBinder :: GenericBinder binder expr -> expr -> GenericArg expr
 argFromBinder (Binder p _ v r _ _) = Arg p v r