--- a/vehicle-syntax/src/Vehicle/Syntax/AST/Binder.hs
+++ b/vehicle-syntax/src/Vehicle/Syntax/AST/Binder.hs
@@ -4,7 +4,6 @@ import Control.DeepSeq (NFData)
 import Data.Aeson (FromJSON, ToJSON)
 import Data.Hashable (Hashable)
 import GHC.Generics (Generic)
-
 import Vehicle.Syntax.AST.Name
 import Vehicle.Syntax.AST.Provenance
 import Vehicle.Syntax.AST.Relevance
@@ -13,20 +12,22 @@ import Vehicle.Syntax.AST.Visibility
 --------------------------------------------------------------------------------
 -- Binders
 
-
 -- | What form the binder appears in the user expression
 data BinderNamingForm
-  -- | Both name and type appear (e.g. {x : A})
-  = NameAndType
-  -- | Only name appears (e.g. {x})
-  | OnlyName
-  -- | Only type appears (e.g. {{HasEq A}})
-  | OnlyType
+  = -- | Both name and type appear (e.g. {x : A})
+    NameAndType
+  | -- | Only name appears (e.g. {x})
+    OnlyName
+  | -- | Only type appears (e.g. {{HasEq A}})
+    OnlyType
   deriving (Eq, Show, Generic)
 
-instance NFData   BinderNamingForm
-instance ToJSON   BinderNamingForm
+instance NFData BinderNamingForm
+
+instance ToJSON BinderNamingForm
+
 instance FromJSON BinderNamingForm
+
 instance Hashable BinderNamingForm
 
 -- | Indicates whether the binder should be folded into the previous binder
@@ -34,13 +35,17 @@ instance Hashable BinderNamingForm
 type BinderFoldingForm = Bool
 
 data BinderForm = BinderForm
-  { namingForm  :: BinderNamingForm
-  , foldingForm :: BinderFoldingForm
-  } deriving (Eq, Show, Generic)
+  { namingForm :: BinderNamingForm,
+    foldingForm :: BinderFoldingForm
+  }
+  deriving (Eq, Show, Generic)
 
-instance NFData   BinderForm
-instance ToJSON   BinderForm
+instance NFData BinderForm
+
+instance ToJSON BinderForm
+
 instance FromJSON BinderForm
+
 instance Hashable BinderForm
 
 -- | Binder for lambda and let expressions
@@ -50,20 +55,23 @@ instance Hashable BinderForm
 -- manually provided by the user it never needs to be updated after unification
 -- and type-class resolution.
 data GenericBinder binder expr = Binder
-  { binderProvenance     :: Provenance
-  , binderForm           :: BinderForm
-  , binderVisibility     :: Visibility
-  -- ^ The visibility of the binder
-  , binderRelevance      :: Relevance
-  -- ^ The relevancy of the binder
-  , binderRepresentation :: binder
-  -- ^ The representation of the bound variable
-  , binderType           :: expr
-  -- The type of the bound variable
-  } deriving (Eq, Show, Functor, Foldable, Traversable, Generic)
+  { binderProvenance :: Provenance,
+    binderForm :: BinderForm,
+    -- | The visibility of the binder
+    binderVisibility :: Visibility,
+    -- | The relevancy of the binder
+    binderRelevance :: Relevance,
+    -- | The representation of the bound variable
+    binderRepresentation :: binder,
+    binderType :: expr
+    -- The type of the bound variable
+  }
+  deriving (Eq, Show, Functor, Foldable, Traversable, Generic)
 
-instance (NFData   binder, NFData   expr) => NFData   (GenericBinder binder expr)
-instance (ToJSON   binder, ToJSON   expr) => ToJSON   (GenericBinder binder expr)
+instance (NFData binder, NFData expr) => NFData (GenericBinder binder expr)
+
+instance (ToJSON binder, ToJSON expr) => ToJSON (GenericBinder binder expr)
+
 instance (FromJSON binder, FromJSON expr) => FromJSON (GenericBinder binder expr)
 
 instance HasProvenance (GenericBinder binder expr) where
@@ -88,10 +96,10 @@ pattern ImplicitBinder :: Provenance -> binder -> expr -> GenericBinder binder e
 pattern ImplicitBinder p n t <- Binder p _ Implicit Relevant n t
 
 pattern InstanceBinder :: Provenance -> binder -> expr -> GenericBinder binder expr
-pattern InstanceBinder p n t <- Binder p _ Instance Relevant  n t
+pattern InstanceBinder p n t <- Binder p _ Instance Relevant n t
 
 pattern IrrelevantInstanceBinder :: Provenance -> binder -> expr -> GenericBinder binder expr
-pattern IrrelevantInstanceBinder p n t <- Binder p _ Instance Irrelevant  n t
+pattern IrrelevantInstanceBinder p n t <- Binder p _ Instance Irrelevant n t
 
 --------------------------------------------------------------------------------
 -- Helper functions
@@ -111,9 +119,10 @@ mapBinderRep f (Binder p u v r b t) = Binder p u v r (f b) t
 replaceBinderRep :: b -> GenericBinder a e -> GenericBinder b e
 replaceBinderRep b' (Binder p u v r _b t) = Binder p u v r b' t
 
-replaceBinderType :: expr1
-                  -> GenericBinder binder expr2
-                  -> GenericBinder binder expr1
+replaceBinderType ::
+  expr1 ->
+  GenericBinder binder expr2 ->
+  GenericBinder binder expr1
 replaceBinderType e = fmap (const e)
 
 wantsToFold :: GenericBinder binder expr -> Bool