--- a/vehicle-syntax/src/Vehicle/Syntax/AST/Builtin/TypeClass.hs
+++ b/vehicle-syntax/src/Vehicle/Syntax/AST/Builtin/TypeClass.hs
@@ -27,6 +27,7 @@ data TypeClass
   | HasDiv
   | HasNeg
   | HasFold
+  | HasMap
   | HasIf
   | HasQuantifierIn Quantifier
   | -- Literal type-classes
@@ -73,6 +74,7 @@ instance Pretty TypeClass where
     HasMul -> "HasMul"
     HasDiv -> "HasDiv"
     HasNeg -> "HasNeg"
+    HasMap -> "HasMap"
     HasFold -> "HasFold"
     HasQuantifierIn q -> "HasQuantifierIn" <+> pretty q
     HasIf -> "HasIf"