… .github/workflows/build-vehicle.yml
… .github/workflows/build.yml
… .pre-commit-config.yaml
… cabal.project.debug
… cabal.project.ghc-debug.ghc-9.2.4.freeze
… cabal.project.nothunks.ghc-9.2.4.freeze
… scripts/stylish-haskell
… vehicle-syntax/src/Vehicle/Syntax/AST/Builtin/Core.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Builtin/Linearity.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Builtin/Polarity.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Decl.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Expr.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Instances/NoThunks.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Meta.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Name.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Prog.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Provenance.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Relevance.hs
… vehicle-syntax/src/Vehicle/Syntax/AST/Visibility.hs
… vehicle-syntax/src/Vehicle/Syntax/BNFC/Delaborate/External.hs
… vehicle-syntax/src/Vehicle/Syntax/BNFC/Delaborate/Internal.hs
… vehicle-syntax/src/Vehicle/Syntax/BNFC/Elaborate/External.hs
… vehicle-syntax/src/Vehicle/Syntax/BNFC/Elaborate/Internal.hs
… vehicle-syntax/src/Vehicle/Syntax/BNFC/Utils.hs
… vehicle-syntax/src/Vehicle/Syntax/Debug.hs
… vehicle-syntax/src/Vehicle/Syntax/Parse.hs
… vehicle-syntax/src/Vehicle/Syntax/Parse/Error.hs
… vehicle-syntax/src/Vehicle/Syntax/Parse/Token.hs
… vehicle-syntax/src/Vehicle/Syntax/Prelude.hs
… vehicle-syntax/src/Vehicle/Syntax/Print.hs