After switching to ormolu and adding HasMap:
```
Benchmark 1: cabal test all
  Time (mean ± σ):      9.901 s ±  0.083 s    [User: 27.202 s, System: 5.107 s]
  Range (min … max):    9.747 s …  9.995 s    10 runs
```