@wenkokke 2022-12-17
Fix tests (77ce5b702503aede2dce5d8261f3738c9d45a554) introduces a performance regression:
```
Benchmark 1: cabal test all
  Time (mean ± σ):     20.536 s ±  0.076 s    [User: 17.196 s, System: 3.560 s]
  Range (min … max):   20.438 s … 20.694 s    10 runs
```