@wenkokke 2022-12-19
Removed 77ce5b702503aede2dce5d8261f3738c9d45a554, added StrictData:
```
Benchmark 1: cabal test all
  Time (mean ± σ):     13.893 s ±  2.906 s    [User: 26.765 s, System: 4.994 s]
  Range (min … max):    9.563 s … 18.417 s    10 runs
```