@wenkokke 2022-12-19
Undo StrictData:
```
Benchmark 1: cabal test all
  Time (mean ± σ):     11.624 s ±  1.647 s    [User: 27.393 s, System: 4.966 s]
  Range (min … max):    9.748 s … 14.081 s    10 runs
```